feat: extract accounts data hooks into features/accounts/hooks/
Create service-aligned data hooks for the accounts feature:
- useAccountsTable: returns queryKey + queryFn for DataTable integration
- useAccountDetail: fetches a single account by ID
- useAccountPostings: fetches ledger postings for an account
- useAccountLiens: fetches active amount blocks for an account

Update AccountsPage and AccountDetailPage to use the new hooks instead
of inline useQuery calls. Also add AccountStatus enum stub to vitest
config so tests resolve the enum correctly without generated proto files.
