+++ frontend/src/features/accounts/hooks/index.ts
@@ -1,0 +1,1 @@
+export { useAccountsTable, useAccountDetail, useAccountPostings, useAccountLiens } from './use-accounts'
