+++ frontend/src/features/accounts/index.ts
@@ -1,3 +1,4 @@
 // Public API for accounts module
 export { AccountsPage } from './pages/index'
 export { AccountDetailPage } from './pages/[accountId]'
+export { useAccountsTable, useAccountDetail, useAccountPostings, useAccountLiens } from './hooks'
