+++ frontend/src/features/accounts/pages/index.tsx
@@ -2,16 +2,13 @@ import * as React from 'react'
 import { useNavigate } from 'react-router-dom'
 import type { ColumnDef } from '@tanstack/react-table'
 import { DataTable } from '@/shared/data-table'
-import type { DataTableQueryParams, DataTableResult } from '@/shared/data-table'
 import { StatusBadge } from '@/shared/status-badge'
 import { TimeDisplay } from '@/shared/time-display'
 import { Button } from '@/components/ui/button'
-import { useApiClients } from '@/api/context'
-import { useTenantContext } from '@/contexts/tenant-context'
-import { tenantKeys } from '@/lib/query-keys'
 import { AccountStatus } from '@/api/gen/meridian/current_account/v1/current_account_pb'
 import { CreateAccountDialog } from './create-account-dialog'
 import type { CurrentAccount } from './types'
+import { useAccountsTable } from '../hooks'
 
 const STATUS_OPTIONS = [
   { label: 'Active', value: String(AccountStatus.ACTIVE) },
@@ -19,17 +16,10 @@ const STATUS_OPTIONS = [
   { label: 'Closed', value: String(AccountStatus.CLOSED) },
 ]
 
-const ACCOUNT_STATUS_NAMES: Record<number, string> = {
-  [AccountStatus.ACTIVE]: 'ACTIVE',
-  [AccountStatus.FROZEN]: 'FROZEN',
-  [AccountStatus.CLOSED]: 'CLOSED',
-}
-
 export function AccountsPage() {
   const navigate = useNavigate()
-  const { tenantSlug } = useTenantContext()
-  const clients = useApiClients()
   const [createOpen, setCreateOpen] = React.useState(false)
+  const { queryKey, queryFn } = useAccountsTable()
 
   const columns: ColumnDef<CurrentAccount>[] = [
     {
@@ -55,44 +45,6 @@ export function AccountsPage() {
       cell: ({ row }) => <TimeDisplay timestamp={row.original.createdAt} format="relative" />,
     },
   ]
-
-  const queryKey = React.useMemo(
-    () => tenantKeys.accounts(tenantSlug ?? ''),
-    [tenantSlug],
-  )
-
-  const queryFn = React.useCallback(
-    async (params: DataTableQueryParams): Promise<DataTableResult<CurrentAccount>> => {
-      if (!tenantSlug) return { items: [] }
-
-      const statusFilter = params.filters?.status
-      const parsedStatus =
-        statusFilter !== undefined && statusFilter !== '' ? Number(statusFilter) : undefined
-      const response = await clients.currentAccount.listCurrentAccounts({
-        pageSize: params.pageSize,
-        pageToken: params.pageToken ?? '',
-        ...(parsedStatus !== undefined && Number.isFinite(parsedStatus)
-          ? { status: parsedStatus as AccountStatus }
-          : {}),
-      })
-
-      const accounts: CurrentAccount[] = (response.accounts ?? []).map((a) => ({
-        accountId: a.accountId ?? '',
-        externalReference: a.externalIdentifier ?? '',
-        status: (ACCOUNT_STATUS_NAMES[a.accountStatus] ?? String(a.accountStatus)) as CurrentAccount['status'],
-        instrumentCode: a.instrumentCode || '',
-        availableBalance: '', // fetched separately via PositionKeeping; not in ListCurrentAccounts response
-        createdAt: a.createdAt ?? undefined,
-        updatedAt: a.updatedAt ?? undefined,
-      }))
-
-      return {
-        items: accounts,
-        nextPageToken: response.nextPageToken || undefined,
-      }
-    },
-    [tenantSlug, clients],
-  )
 
   return (
     <div className="p-6">
